api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Fèfèn — chatbot TF-IDF + RAG
scikit-learn>=1.4.0
numpy>=1.26.0
joblib>=1.3.0
huggingface_hub>=0.22.0

# Auth JWT
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Tests
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
